+++ src/jk_utils/datetime/__init__.py
@@ -1,7 +1,4 @@
-﻿
 
-
-__version__ = "0.2022.2.6.1"
 
 
 from .D import D