+++ src/jk_utils/tokenizer2/__init__.py
@@ -1,6 +1,4 @@
-﻿
 
-__version__ = "0.2022.2.6.1"
 
 
 from .Token import Token